Store keyset count
