+++ crates/bcr-wallet-core/src/db/mod.rs
@@ -1,6 +1,4 @@
-mod manager;
 pub mod rexie;
 mod types;
 
-pub use manager::Manager;
 pub use types::*;
