+++ crates/bcr-wallet-core/src/db/rexie/constants.rs
@@ -1,0 +1,1 @@
+pub const KEYSET_COUNTER: &str = "KEYSET_COUNTER";
